splash arch updated to mvvm
