+++ lib/feature/splash/model/splash.dart
@@ -1,0 +1,11 @@
+/* import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'splash.freezed.dart';
+
+@freezed
+class Splash with _$Splash {
+  const factory Splash({
+    required bool isConnected,
+  }) = _Splash;
+}
+ */
