--- lib/feature/splash/presentation/splash_view.dart
@@ -1,30 +1,0 @@
-import 'package:auto_route/auto_route.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mdp2/feature/splash/presentation/splash_mixin.dart';
-
-@RoutePage()
-class SplashView extends ConsumerStatefulWidget {
-  const SplashView({super.key});
-
-  @override
-  ConsumerState<ConsumerStatefulWidget> createState() => _SplashViewState();
-}
-
-class _SplashViewState extends ConsumerState<SplashView> with SplashMixin {
-  @override
-  Widget build(BuildContext context) {
-    listenAndNavigate();
-    return const Scaffold(
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            CircularProgressIndicator(),
-            Text('Splash View'),
-          ],
-        ),
-      ),
-    );
-  }
-}
