+++ lib/feature/splash/view_model/splash_view_model.dart
@@ -1,0 +1,27 @@
+import 'package:connectivity_plus/connectivity_plus.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+part 'splash_view_model.g.dart';
+
+@riverpod
+class SplashViewModel extends _$SplashViewModel {
+  @override
+  Future<bool> build() async {
+    return checkInternetConnection();
+  }
+
+  Future<bool> checkInternetConnection() async {
+    final connectivityResult = await Connectivity().checkConnectivity();
+
+    await _delay();
+    if (connectivityResult.contains(ConnectivityResult.none)) {
+      return false;
+    } else {
+      return true;
+    }
+  }
+
+  Future<void> _delay() async {
+    await Future<void>.delayed(const Duration(seconds: 2));
+  }
+}
